update unit test

+++ backend/Ecommerce/src/main/java/rookies/training/service/impl/CategoryServiceImpl.java
@@ -1,9 +1,11 @@
 package rookies.training.service.impl;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import com.google.gson.Gson;
+import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,7 @@ import rookies.training.repository.CategoryRepository;
 import rookies.training.service.CategoryService;
 
 @Service
+@AllArgsConstructor
 public class CategoryServiceImpl implements CategoryService{
 	
 	@Autowired
@@ -23,7 +26,11 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Autowired
 	ModelMapper modelMapper;
-	
+
+//	public CategoryServiceImpl(CategoryRepository categoryRepository, ModelMapper modelMapper) {
+//		this.modelMapper =
+//	}
+
 	@Override
 	public List<CategoryDTO> getAllCategories() {
 		return categoryRepository.findAll().stream().map(category -> modelMapper.map(category,CategoryDTO.class)).collect(Collectors.toList());
@@ -31,22 +38,24 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Override
 	public CategoryDTO getCategoryById(Long id) {
-		Category category=categoryRepository.findById(id).get();
-		return modelMapper.map(category,CategoryDTO.class);
+		Optional<Category> optionCategory=categoryRepository.findById(id);
+		if (optionCategory.isEmpty())	return null;
+		return modelMapper.map(optionCategory.get(),CategoryDTO.class);
 	}
 
 	@Override
 	public CategoryDTO createCategory(CategoryDTO categoryDTO) {
 		Category category=modelMapper.map(categoryDTO,Category.class);
-		categoryRepository.saveAndFlush(category);
+		category=categoryRepository.save(category);
 		CategoryDTO res= modelMapper.map(category,CategoryDTO.class);
 		return res;
 	}
 
 	@Override
 	public String updateCategory(Long id,Category category) {
-		Category category2=categoryRepository.findById(id).get();
-		if (category2==null)	return "category is not exist";
+		Optional<Category> optionCategory2=categoryRepository.findById(id);
+		if (optionCategory2.isEmpty())	return "category is not exist";
+		Category category2=optionCategory2.get();
 		category2.setName(category.getName());
 		category2.setUrl(category.getUrl());
 		category2.setUrl(category.getLogoUrl());
@@ -57,13 +66,15 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Override
 	public void deleteCategoryById(Long id) {
+		Optional<Category> optionalCategory=categoryRepository.findById(id);
+		if (optionalCategory.isEmpty())	return;
 		categoryRepository.deleteById(id);
 	}
 
 	@Override
 	public List<ProductDTO> getProductsByCategory(Long id) {
-		Category category=categoryRepository.findById(id).get();
-		if (category==null)	return null;
-		return category.getListProduct().stream().map((product)->(modelMapper.map(product,ProductDTO.class))).collect(Collectors.toList());
+		Optional<Category> optionalCategory=categoryRepository.findById(id);
+		if (optionalCategory.isEmpty())	return null;
+		return optionalCategory.get().getListProduct().stream().map((product)->(modelMapper.map(product,ProductDTO.class))).collect(Collectors.toList());
 	}
 }
